Refactor profile slab
